feat: inline type-hints + split into multiple files
