+++ src/nntp/__init__.py
@@ -1,0 +1,22 @@
+from nntp._core import NNTP, NNTP_SSL
+from nntp._exceptions import (
+    NNTPDataError,
+    NNTPError,
+    NNTPPermanentError,
+    NNTPProtocolError,
+    NNTPReplyError,
+    NNTPTemporaryError,
+)
+from nntp._helpers import decode_header
+
+__all__ = [
+    "NNTP",
+    "NNTPError",
+    "NNTPReplyError",
+    "NNTPTemporaryError",
+    "NNTPPermanentError",
+    "NNTPProtocolError",
+    "NNTPDataError",
+    "decode_header",
+    "NNTP_SSL",
+]
